requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
Crawl4AI==0.6.3
cryptography==45.0.3
cssselect==1.3.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.1
email_validator==2.2.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
graphviz==0.20.3
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio<1.73
grpcio-status<1.73
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanize==4.12.3
idna==3.10
imagekitio==3.2.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
litellm==1.72.2
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.3
mdurl==0.1.2
multidict==6.4.4
nltk==3.9.1
numpy==2.3.0
openai==1.84.0
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
passlib==1.7.4
pillow==10.4.0
playwright==1.52.0
propcache==0.3.1
proto-plus==1.26.1
protobuf>=4.25,<5
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==0.10.1
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
sib-api-v3-sdk==7.6.0
firebase-admin
google-search-results
markdownify
pydantic-ai
google-generativeai>=0.6.0
azure-identity
azure-ai-ml
replicate
google-auth
passlib[bcrypt]
tiktoken==0.9.0